requirements.txt
transformers==4.30.0
tqdm==4.65.0
Pillow==9.4.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
chardet==5.1.0